got eye beam GCDs to stop overlapping with pure black magic
