+++ src/Parser/DemonHunter/Havoc/Modules/Core/GlobalCooldown.js
@@ -5,7 +5,15 @@ import CoreGlobalCooldown from 'Parser/Core/Modules/GlobalCooldown';
  * Eye Beam's cast event is actually a beginchannel event, so it shouldn't add the GCD as active time.
  */
 class GlobalCooldown extends CoreGlobalCooldown {
-  on_byPlayer_applybuff(event) {
+  // on_byPlayer_applybuff(event) {
+  //   if (event.ability.guid === SPELLS.EYE_BEAM.id) {
+  //     // This GCD gets handled by the `beginchannel` event
+  //     return;
+  //   }
+  //   super.on_byPlayer_cast(event);
+  // }
+
+  on_byPlayer_cast(event) {
     if (event.ability.guid === SPELLS.EYE_BEAM.id) {
       // This GCD gets handled by the `beginchannel` event
       return;
